algunos cambios

+++ src/main/kotlin/cu/suitetecsa/sdk/nauta/domain/model/ConnectionsSummary.kt
@@ -3,7 +3,7 @@ package cu.suitetecsa.sdk.nauta.domain.model
 data class ConnectionsSummary(
     val count: Int,
     val yearMonthSelected: String,
-    val totalTime: Int,
+    val totalTime: Long,
     val totalImport: Float,
     val uploaded: Double,
     val downloaded: Double,
